fix(issues): usa BigInteger para IDs globais do GitHub

+++ backend/app/models/issue.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import uuid
 from datetime import datetime
 
-from sqlalchemy import Boolean, DateTime, ForeignKey, Index, Integer, String, Text, UniqueConstraint
+from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Index, Integer, String, Text, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.base import Base, UUIDPrimaryKeyMixin
@@ -20,7 +20,7 @@ class Issue(UUIDPrimaryKeyMixin, Base):
     repository_id: Mapped[uuid.UUID] = mapped_column(
         ForeignKey("repositories.id", ondelete="CASCADE"), nullable=False, index=True
     )
-    github_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    github_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     number: Mapped[int] = mapped_column(Integer, nullable=False)
     title: Mapped[str] = mapped_column(String(1000), nullable=False)
     state: Mapped[str] = mapped_column(String(30), nullable=False)
